Rename Microsoft.AspNet.HttpFeature to Microsoft.AspNet.Http.Interfaces
