+++ src/Microsoft.AspNet.Http.Core/DefaultHttpResponse.cs
@@ -12,8 +12,8 @@ using Microsoft.AspNet.Http;
 using Microsoft.AspNet.Http.Infrastructure;
 using Microsoft.AspNet.Http.Security;
 using Microsoft.AspNet.FeatureModel;
-using Microsoft.AspNet.HttpFeature;
-using Microsoft.AspNet.HttpFeature.Security;
+using Microsoft.AspNet.Http.Interfaces;
+using Microsoft.AspNet.Http.Interfaces.Security;
 using Microsoft.AspNet.Http.Core.Collections;
 using Microsoft.AspNet.Http.Core.Infrastructure;
 using Microsoft.AspNet.Http.Core.Security;
