+++ src/Microsoft.AspNet.Http.Core/ResponseCookiesFeature.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.AspNet.Http;
 using Microsoft.AspNet.FeatureModel;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 using Microsoft.AspNet.Http.Core.Collections;
 using Microsoft.AspNet.Http.Core.Infrastructure;
