+++ src/Microsoft.AspNet.Http.Core/WebSocketAcceptContext.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 
 namespace Microsoft.AspNet.Http.Core
 {
