+++ src/Microsoft.AspNet.Http.Core/project.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "Microsoft.AspNet.FeatureModel": "1.0.0-*",
     "Microsoft.AspNet.Http": "1.0.0-*",
-    "Microsoft.AspNet.HttpFeature": "1.0.0-*",
+    "Microsoft.AspNet.Http.Interfaces": "1.0.0-*",
     "Microsoft.AspNet.WebUtilities": "1.0.0-*",
     "Microsoft.Net.Http.Headers": "1.0.0-*"
