+++ src/Microsoft.AspNet.Http.Interfaces/AssemblyNeutralAttribute.cs
