+++ src/Microsoft.AspNet.Http.Interfaces/IHttpBufferingFeature.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
     public interface IHttpBufferingFeature
