+++ src/Microsoft.AspNet.Http.Interfaces/IHttpClientCertificateFeature.cs
@@ -6,7 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
     public interface IHttpClientCertificateFeature
