+++ src/Microsoft.AspNet.Http.Interfaces/IHttpConnectionFeature.cs
@@ -4,7 +4,7 @@
 using System.Net;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
     public interface IHttpConnectionFeature
