+++ src/Microsoft.AspNet.Http.Interfaces/IHttpRequestFeature.cs
@@ -6,16 +6,18 @@ using System.Collections.Generic;
 using System.IO;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
-    public interface IHttpResponseFeature
+    public interface IHttpRequestFeature
     {
-        int StatusCode { get; set; }
-        string ReasonPhrase { get; set; }
+        string Protocol { get; set; }
+        string Scheme { get; set; }
+        string Method { get; set; }
+        string PathBase { get; set; }
+        string Path { get; set; }
+        string QueryString { get; set; }
         IDictionary<string, string[]> Headers { get; set; }
         Stream Body { get; set; }
-        bool HeadersSent { get; }
-        void OnSendingHeaders(Action<object> callback, object state);
     }
 }
