+++ src/Microsoft.AspNet.Http.Interfaces/IHttpRequestLifetimeFeature.cs
@@ -1,17 +1,15 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Net.WebSockets;
-using System.Threading.Tasks;
+using System.Threading;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
-    public interface IHttpWebSocketFeature
+    public interface IHttpRequestLifetimeFeature
     {
-        bool IsWebSocketRequest { get; }
-
-        Task<WebSocket> AcceptAsync(IWebSocketAcceptContext context);
+        CancellationToken RequestAborted { get; }
+        void Abort();
     }
 }
