+++ src/Microsoft.AspNet.Http.Interfaces/IHttpWebSocketFeature.cs
@@ -1,15 +1,17 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Threading;
+using System.Net.WebSockets;
+using System.Threading.Tasks;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
-    public interface IHttpRequestLifetimeFeature
+    public interface IHttpWebSocketFeature
     {
-        CancellationToken RequestAborted { get; }
-        void Abort();
+        bool IsWebSocketRequest { get; }
+
+        Task<WebSocket> AcceptAsync(IWebSocketAcceptContext context);
     }
 }
