+++ src/Microsoft.AspNet.Http.Interfaces/ISession.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
     public interface ISession
