+++ src/Microsoft.AspNet.Http.Interfaces/ISessionFactory.cs
@@ -1,13 +1,16 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
-    public interface IWebSocketAcceptContext
+    public interface ISessionFactory
     {
-        string SubProtocol { get; set; }
+        bool IsAvailable { get; }
+
+        ISession Create();
     }
 }
