+++ src/Microsoft.AspNet.Http.Interfaces/IWebSocketAcceptContext.cs
@@ -1,16 +1,13 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.IO;
-using System.Threading.Tasks;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature
+namespace Microsoft.AspNet.Http.Interfaces
 {
     [AssemblyNeutral]
-    public interface IHttpUpgradeFeature
+    public interface IWebSocketAcceptContext
     {
-        bool IsUpgradableRequest { get; }
-        Task<Stream> UpgradeAsync();
+        string SubProtocol { get; set; }
     }
 }
