+++ src/Microsoft.AspNet.Http.Interfaces/Security/IAuthenticateContext.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Security.Claims;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature.Security
+namespace Microsoft.AspNet.Http.Interfaces.Security
 {
     [AssemblyNeutral]
     public interface IAuthenticateContext
