+++ src/Microsoft.AspNet.Http.Interfaces/Security/IChallengeContext.cs
@@ -4,13 +4,14 @@
 using System.Collections.Generic;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature.Security
+namespace Microsoft.AspNet.Http.Interfaces.Security
 {
     [AssemblyNeutral]
-    public interface ISignOutContext 
+    public interface IChallengeContext
     {
-        IEnumerable<string> AuthenticationTypes { get; }
+        IEnumerable<string> AuthenticationTypes {get;}
+        IDictionary<string,string> Properties {get;}
 
-        void Accept(string authenticationType, IDictionary<string, object> description);
+        void Accept(string authenticationType, IDictionary<string,object> description);
     }
 }
