+++ src/Microsoft.AspNet.Http.Interfaces/Security/IHttpAuthenticationFeature.cs
@@ -4,7 +4,7 @@
 using System.Security.Claims;
 using Microsoft.Framework.Runtime;
 
-namespace Microsoft.AspNet.HttpFeature.Security
+namespace Microsoft.AspNet.Http.Interfaces.Security
 {
     [AssemblyNeutral]
     public interface IHttpAuthenticationFeature
