+++ src/Microsoft.AspNet.Http.Interfaces/project.json
