+++ src/Microsoft.AspNet.Owin/WebSockets/OwinWebSocketAcceptAdapter.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Net.WebSockets;
 using System.Threading.Tasks;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 
 namespace Microsoft.AspNet.Owin
 {
