+++ src/Microsoft.AspNet.Owin/WebSockets/OwinWebSocketAcceptContext.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 
 namespace Microsoft.AspNet.Owin
 {
