+++ src/Microsoft.AspNet.Owin/WebSockets/WebSocketAcceptAdapter.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Net.WebSockets;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 
 namespace Microsoft.AspNet.Owin
 {
