+++ src/Microsoft.AspNet.Owin/project.json
@@ -5,7 +5,7 @@
         "Microsoft.AspNet.Http": "1.0.0-*",
         "Microsoft.AspNet.FeatureModel": "1.0.0-*",
         "Microsoft.AspNet.Http.Core": "1.0.0-*",
-        "Microsoft.AspNet.HttpFeature": { "version": "1.0.0-*", "type": "build" }
+        "Microsoft.AspNet.Http.Interfaces": { "version": "1.0.0-*", "type": "build" }
     },
     "frameworks": {
         "aspnet50": { },
