+++ test/Microsoft.AspNet.Http.Tests/MapPathMiddlewareTests.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.AspNet.Http;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 using Microsoft.AspNet.Http.Core;
 using Shouldly;
 using Xunit;
