+++ test/Microsoft.AspNet.Http.Tests/MapPredicateMiddlewareTests.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNet.Http;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 using Microsoft.AspNet.Http.Core;
 using Xunit;
