+++ test/Microsoft.AspNet.Http.Tests/project.json
@@ -1,7 +1,7 @@
 {
   "dependencies": {
     "Microsoft.AspNet.Http": "1.0.0-*",
-    "Microsoft.AspNet.HttpFeature": "1.0.0-*",
+    "Microsoft.AspNet.Http.Interfaces": "1.0.0-*",
     "Microsoft.AspNet.Http.Core": "1.0.0-*",
     "Microsoft.AspNet.Testing": "1.0.0-*",
     "xunit.runner.kre": "1.0.0-*"
