+++ test/Microsoft.AspNet.Owin.Tests/OwinFeatureCollectionTests.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNet.FeatureModel;
-using Microsoft.AspNet.HttpFeature;
+using Microsoft.AspNet.Http.Interfaces;
 using Xunit;
 
 namespace Microsoft.AspNet.Owin
